+++ src/main/java/duke/exception/InvalidIndex.java
@@ -1,0 +1,2 @@
+package duke;public class InvalidIndex {
+}
